requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
keras==2.15.0
tensorflow==2.15.0
numpy==1.25.2
Pillow==10.0.0
requests==2.31.0